LayoutTests/http/tests/frame-throttling/timer-throttle-in-cross-origin-subframe-expected.txt

Tests that DOM Timers are throttled in subframes that are cross-origin, and not in same-origin frames

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Checking that DOM Timers are throttled in cross origin frame
PASS throttledState['cross'] is "true"
PASS throttledState['same'] is "false"
Interacted with cross-origin frame
Interacted with same-origin frame
PASS throttledState['cross'] is "false"
PASS throttledState['same'] is "false"
PASS successfullyParsed is true

TEST COMPLETE
